Add authorization and error handling

+++ backend/routes/posts.js
@@ -1,6 +1,8 @@
 const express = require("express");
-const Post = require("../models/post");
 const multer = require("multer");
+
+const Post = require("../models/post");
+const checkAuth = require("../middleware/check-auth");
 
 const router = express.Router();
 
@@ -26,12 +28,13 @@ const storage = multer.diskStorage({
   }
 });
 
-router.post("", multer({storage: storage}).single("image"), (req, res, next) => {
+router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
   const url = req.protocol + '://' + req.get("host");
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(createdPost => {
     res.status(201).json({
@@ -41,11 +44,17 @@ router.post("", multer({storage: storage}).single("image"), (req, res, next) => 
         id: createdPost._id
       }
     });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Creating a post failed!"
+    })
   });
 });
 
 router.put(
   "/:id",
+  checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
@@ -57,11 +66,20 @@ router.put(
       _id: req.body.id,
       title: req.body.title,
       content: req.body.content,
-      imagePath: imagePath
+      imagePath: imagePath,
+      creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({_id: req.params.id}, post).then( result => {
-      res.status(200).json({message: "Update successful!"});
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then( result => {
+      if (result.nModified > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({message: "Not Authorized!"});
+      }
+    })
+    .catch(error => {
+      res.status(500).json({
+        message: "Couldn't update post!"
+      })
     });
   }
 );
@@ -87,6 +105,11 @@ router.get("", (req, res, next) => {
       posts: fetchedPosts,
       maxPosts: count
     });
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Fetching posts failed!"
+    });
   });
 });
 
@@ -97,13 +120,25 @@ router.get("/:id", (req, res, next) => {
     } else {
       res.status(404).json({message: "Post not found!"});
     }
+  }).catch(error => {
+    res.status(500).json({
+      message: "Fetching a post failed!"
+    });
   });
 });
 
-router.delete("/:id", (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+router.delete("/:id", checkAuth, (req, res, next) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({message: "Deletion successful!"});
+    } else {
+      res.status(401).json({message: "Not Authorized!"});
+    }
+  })
+  .catch(error => {
+    res.status(500).json({
+      message: "Fetching posts failed!"
+    });
   });
 });
